"Refactored server.py and optimized registration time in create_mail.py
Removed unused code and unused imports from server.py, not only to clean up the file but also to progress towards a more efficient and focused codebase. Additionally, replaced the function to create and register proper threads with the threading module for efficiency.

In create_mail.py, decreased the webdriver's wait time from 10 to 5 seconds to speed up the registration process. This alteration takes into account the improved loading time of the pages, aiming to provide a faster and smoother user experience."