"Refactored server.py into an AutoReferral class
The program's structure has been modified to improve maintainability and scalability. The initial server.py, which contained global threads and functions, was removed and replaced with an AutoReferral class in the autoreferral.py file. Now specific parameters like the referral link and number of registrations can be set when creating an instance of this class. Functionality was also transferred from run.py to this AutoRefreferral class to consolidate the application's functionalities within a single object-oriented design."