"Improved input validation in run.py script
Refactored to include better input validation in the script. Instead of simply asking for inputs, the code now contains loop structures that only break if valid inputs are received. Added a 'ValueError' exception for number of registrations input to handle non-integer inputs and a conditional check to make sure the input is not less than or equal to zero. Also optimized the input process for showing the registration process, ensuring that user can only input 'y' or 'n', prompting them to choose a valid option if anything else is entered.